extract fields into individual module
