+++ lib/model.js
@@ -4,47 +4,11 @@ const pluralize = require('pluralize');
 const isObject = require('is-plain-obj');
 const result = require('lodash.result');
 const arrify = require('arrify');
-const get = require('get-value');
-const set = require('set-value');
 
 const queryMethods = require('../util/query-methods');
 const flatten = require('../util/join-obj-keys');
+const Fields = require('./fields');
 const Query = require('./query');
-
-class Fields {
-	constructor(fields = {}) {
-		this.fields = {};
-		this.set(fields);
-	}
-
-	get(key) {
-		return key ? get(this.fields, key) : this.fields;
-	}
-
-	set(key, value) {
-		if (isObject(key)) {
-			const obj = flatten(key);
-
-			Object.keys(obj).forEach(key => {
-				this.set(key, obj[key]);
-			});
-
-			return;
-		}
-
-		set(this.fields, key, value);
-	}
-
-	unset(key = []) {
-		arrify(key).forEach(key => {
-			delete this.fields[key];
-		});
-	}
-
-	clear() {
-		this.fields = {};
-	}
-}
 
 class Model {
 	constructor(fields = {}) {
